Move TB test to tb_test.go
